backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.1
gunicorn==20.1.0
requests==2.28.2
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
joblib==1.2.0
matplotlib==3.7.1
xgboost==1.7.5
qrcode[pil]==7.4.2